+++ app/Http/Controllers/FrenosController.php
@@ -3,19 +3,23 @@
 namespace App\Http\Controllers;
 
 use App\Models\Frenos;
+use App\Models\Tarea;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 // TODO: Borrar las respuestas de error para que no se exponga información sensible en producción. Se considerará agregar visualización detallada de productos si aplica. Se tiene que agregar la lógica para permitir el uso de imágenes
 
 class FrenosController extends Controller {
+    use AuthorizesRequests;
     /**
      * Listado de todos los frenos paginados
      */
     public function index(): JsonResponse {
         try {
             $frenos = Frenos::with('tarea')->paginate(10);
-            
+
             return response()->json([
                 'status'  => true,
                 'data'    => $frenos,
@@ -43,6 +47,9 @@ class FrenosController extends Controller {
         ]);
 
         try {
+            $tarea = Tarea::findOrFail($validador['tarea_id']);
+            $this->authorize('checar-id-mecanico', $tarea);
+
             $nuevo_freno = Frenos::create([
                 'tarea_id'        => $validador['tarea_id'],
                 'delanteros'      => $validador['delanteros'],
@@ -79,6 +86,12 @@ class FrenosController extends Controller {
 
         try {
             $freno = Frenos::findOrFail($id);
+            $tarea = $freno->tarea;
+
+            if (!Gate::allows('checar-id-mecanico', $tarea)){
+                return response()->json(['error' => 'Accion no autorizada'], 403);
+            }
+
             $freno->update($validador);
 
             return response()->json([
@@ -101,6 +114,12 @@ class FrenosController extends Controller {
     public function destroy(string $id): JsonResponse {
         try {
             $freno = Frenos::findOrFail($id);
+            $tarea = $freno->tarea;
+
+            if (!Gate::allows('checar-id-mecanico', $tarea)){
+                return response()->json(['error' => 'Accion no autorizada'], 403);
+            }
+
             $freno->delete();
 
             return response()->json(null, 204);
