+++ app/Http/Controllers/TrenDelanteroController.php
@@ -2,20 +2,24 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Tarea;
 use App\Models\TrenDelantero;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 // TODO: Borrar las respuestas de error para que no se exponga información sensible en producción. Se considerará agregar visualización detallada de productos si aplica. Se tiene que agregar la lógica para permitir el uso de imágenes
 
 class TrenDelanteroController extends Controller {
+    use AuthorizesRequests;
     /**
      * Listado de todos los trenes delanteros paginados
      */
     public function index(): JsonResponse {
         try {
             $trenes_delanteros = TrenDelantero::with('tarea')->paginate(10);
-            
+
             return response()->json([
                 'status'  => true,
                 'data'    => $trenes_delanteros,
@@ -50,6 +54,10 @@ class TrenDelanteroController extends Controller {
         ]);
 
         try {
+
+            $tarea = Tarea::findOrFail($validador['tarea_id']);
+            $this->authorize('checar-id-mecanico', $tarea);
+
             $nuevo_tren_delantero = TrenDelantero::create([
                 'tarea_id'       => $validador['tarea_id'],
                 'conv'           => $validador['conv'],
@@ -100,6 +108,12 @@ class TrenDelanteroController extends Controller {
 
         try {
             $tren_delantero = TrenDelantero::findOrFail($id);
+            $tarea = $tren_delantero->tarea;
+
+            if (!Gate::allows('checar-id-mecanico', $tarea)){
+                return response()->json(['error' => 'Accion no autorizada'], 403);
+            }
+
             $tren_delantero->update($validador);
 
             return response()->json([
@@ -122,6 +136,12 @@ class TrenDelanteroController extends Controller {
     public function destroy(string $id): JsonResponse {
         try {
             $tren_delantero = TrenDelantero::findOrFail($id);
+            $tarea = $tren_delantero->tarea;
+
+            if (!Gate::allows('checar-id-mecanico', $tarea)){
+                return response()->json(['error' => 'Accion no autorizada'], 403);
+            }
+
             $tren_delantero->delete();
 
             return response()->json(null, 204);
